feat(test-audit): extract shared FakeClient infrastructure from consolidation_review tests (Closes #859)
Move _FakeClient, _FakeTableQuery, _FakeRPC, _FakeResp, and _filter_val
into tests/supabase_fake.py as a reusable module. Also fix a pre-existing
environment issue where the local supabase/ CLI directory shadows the real
SDK, causing stub setup to silently fail.

test_consolidation_review.py now imports FakeClient/filter_val from the
shared module, reducing 121 lines of inline fakes to a 1-line import.
